hive refactor
