+++ packages/crdt_lf_hive/lib/src/adapters/ids.dart
@@ -1,20 +1,5 @@
-/// hlc adapter id
-const kHybridLogicalClockAdapter = 101;
-
 /// change adapter id
 const kChangeAdapter = 103;
 
-/// peer id adapter id
-const kPeerIdAdapter = 104;
-
 /// snapshot adapter id
 const kSnapshotAdapter = 105;
-
-/// version vector adapter id
-const kVersionVectorAdapter = 106;
-
-/// fugue value node adapter id
-const kFugueValueNodeAdapter = 107;
-
-/// fugue element id adapter id
-const kFugueElementIDAdapter = 108;
